+++ src/main/java/store/lastdance/dto/analysis/SaveAnalysisResultRequestDTO.java
@@ -1,9 +1,7 @@
-package store.lastdance.dto.expense;
+package store.lastdance.dto.analysis;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
-import store.lastdance.dto.analysis.AnalyzeExpenseRequestDTO;
-import store.lastdance.dto.analysis.AnalyzeExpenseResponseDTO;
 
 public record SaveAnalysisResultRequestDTO(
         @NotNull @Valid AnalyzeExpenseRequestDTO requestDTO,
